requirements.txt
asgiref
Django
pillow
sqlparse
babel
Brotli
clic
dj-database-url
django-ckeditor-5
django-money
django-unfold
gunicorn
h11
packaging
pi
psycopg2-binary
py-moneyed
setuptools
typing_extensions
uvicorn
whitenoise